[Fix] 레시피 재료 개수 타입 변경 integer -> double

+++ src/main/java/com/inmyhand/refrigerator/recipe/domain/dto2/RecipeIngredientEntityDto.java
@@ -21,7 +21,7 @@ public class RecipeIngredientEntityDto implements Serializable {
     private String ingredientGroup;
 
     @Min(0)
-    private Integer ingredientQuantity;
+    private double ingredientQuantity;
 
     private String ingredientUnit;
 }
